+++ template/UtilisateurDetail.php
@@ -124,14 +124,15 @@
 		<?php $this->displayCSRFInput(); ?>
 		<input type='hidden' name='id_u' value='<?php echo $id_u ?>' />
 	
-		<select name='role' class='zselect_role'>
+		<select name='role' class='select2_role'>
 			<option value=''>...</option>
 			<?php foreach($role_authorized as $role_info ): ?>
 				<option value='<?php echo $role_info['role']?>'> <?php echo $role_info['libelle'] ?> </option>
 			<?php endforeach ; ?>
 		</select>
 		
-		<select name='id_e' class='zselect_entite'>
+		<select name='id_e' class='select2_entite'>
+            <option></option>
 			<option value='0'>Entité racine</option>
 			<?php foreach($arbre as $entiteInfo): ?>
 				<option value='<?php echo $entiteInfo['id_e']?>'>
@@ -211,7 +212,8 @@
 		<?php $this->displayCSRFInput(); ?>
 		<input type='hidden' name='id_u' value='<?php echo $id_u ?>' />
 		
-		<select name='id_e' class='zselect_entite'>
+		<select name='id_e' class='select2_entite'>
+            <option></option>
 			<option value='0'>Entité racine</option>
 			<?php foreach($arbre as $entiteInfo): ?>
 				<option value='<?php echo $entiteInfo['id_e']?>'><?php echo $entiteInfo['denomination']?> </option>